requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pyjwt==2.8.0
gunicorn==21.2.0
SQLAlchemy==2.0.28
Werkzeug==2.3.7
email-validator==2.1.0.post1 